+++ examples/getting_started/lib/main.dart
@@ -1,24 +1,28 @@
 // @dart=2.12
-import 'dart:async';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:ff_flutter_client_sdk/CfClient.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:logging/logging.dart';
-
-// The SDK API Key to use for authentication.  Configure it when installing the app by setting FF_API_KEY
-// e.g..
-const apiKey = String.fromEnvironment('FF_API_KEY', defaultValue: '');
 
 const boolFlagName = 'boolflag';
 const stringFlagName = "multivariateflag";
 const numberFlagName = "numberflag";
 const jsonFlagName = "jsonflag";
 
-void main() => runApp(MyApp());
+// The SDK API Key to use for authentication.
+// final provider.UniversalApiKeyProvider apiKeyProvider = kIsWeb ? provider() : MobileApiKeyProvider();
+
+void main() async {
+  await dotenv.load(fileName: ".env");
+  runApp(MyApp());
+}
 
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(title: 'Harness Flutter SDK Getting Started', home: FlagState());
+    return MaterialApp(
+        title: 'Harness Flutter SDK Getting Started', home: FlagState());
   }
 }
 
@@ -28,10 +32,12 @@ class FlagState extends StatefulWidget {
 }
 
 class _FlagState extends State<FlagState> {
-  dynamic _boolFlagValue = false;
-  dynamic _stringFlagValue = "off";
-  dynamic _numberFlagValue = 3;
-  dynamic _jsonFlagValue = {};
+  final Map<String, dynamic> _flagValues = {
+    boolFlagName: null,
+    stringFlagName: null,
+    numberFlagName: null,
+    jsonFlagName: null,
+  };
 
   @override
   void initState() {
@@ -39,101 +45,116 @@ class _FlagState extends State<FlagState> {
 
     // Create Default Configuration for the SDK.  We can use this to disable streaming,
     // change the URL the client connects to etc
-    var conf = CfConfigurationBuilder().setLogLevel(Level.FINE).build();
+    var conf = CfConfigurationBuilder()
+        .setLogLevel(Level.FINE)
+        .setStreamEnabled(true)
+        .setDebugEnabled(true)
+        .build();
 
     // Create a target (different targets can get different results based on rules.  This include a custom attribute 'location')
-    var target = CfTargetBuilder().setIdentifier("fluttersdk").setName("FlutterSDK").build();
+    var target = CfTargetBuilder()
+        .setIdentifier("fluttersdk")
+        .setName("FlutterSDK")
+        .build();
+
+    var apiKey = dotenv.env['FF_API_KEY'];
+
+    if (apiKey == null) {
+      print("API Key missing, existing FF Sample application");
+      return;
+    }
 
     // Init the default instance of the Feature Flag Client
-    CfClient.getInstance().initialize(apiKey, conf, target)
-        .then((value){
-      if (value.success) {
+    CfClient.getInstance().initialize(apiKey, conf, target).then((initResult) {
+      if (initResult.success) {
         print("Successfully initialized client");
 
-        // Evaluate flag and set initial state
-        CfClient.getInstance().boolVariation(boolFlagName, false).then((value) {
-          print("$_boolFlagValue: $value");
-          setState(() {
-            _boolFlagValue = value;
-          });
-        });
-
-        // Evaluate flag and set initial state
-        CfClient.getInstance().jsonVariation(jsonFlagName, {}).then((value) {
-          print("$_jsonFlagValue: $value");
-          setState(() {
-            _jsonFlagValue = value;
-          });
-        });
-
-        // Evaluate flag and set initial state
-        CfClient.getInstance().stringVariation(stringFlagName, "default").then((value) {
-          print("$_stringFlagValue: $value");
-          setState(() {
-            _stringFlagValue = value;
-          });
-        });
-
-        // Evaluate flag and set initial state
-        CfClient.getInstance().numberVariation(numberFlagName, 1).then((value) {
-          print("$_numberFlagValue: $value");
-          setState(() {
-            _numberFlagValue = value;
-          });
-        });
+        // Evaluate flags and set initial state
+        flagVariations();
 
         // Setup Event Handler
         listener(data, eventType) {
-          print("received event: ${eventType.toString()} with Data: ${data.toString()}");
+          print("received event: ${eventType.toString()}");
           switch (eventType) {
             case EventType.EVALUATION_CHANGE:
               String flag = (data as EvaluationResponse).flag;
-              dynamic value = (data as EvaluationResponse).value;
-              switch(flag) {
-                case boolFlagName:
-                  setState(() {
-                    _boolFlagValue = value;
-                  });
-                  break;
-                case stringFlagName:
-                  setState(() {
-                    _stringFlagValue = value;
-                  });
-                  break;
-                case numberFlagName:
-                  setState(() {
-                    _numberFlagValue = value;
-                  });
-                  break;
-                case jsonFlagName:
-                  setState(() {
-                    _jsonFlagValue = value;
-                  });
-                  break;
+
+              if (_flagValues.containsKey(flag)) {
+                setState(() {
+                  print(
+                      "Flag evaluation changed via streaming event: Flag: '$flag', New Evaluation: ${data.value}");
+                  _flagValues[flag] = data.value;
+                });
               }
               break;
 
+            case EventType.EVALUATION_POLLING:
+              List<EvaluationResponse> evals =
+                  (data as List<EvaluationResponse>);
 
+              for (final eval in evals) {
+                if (_flagValues.containsKey(eval.flag)) {
+                  setState(() {
+                    _flagValues[eval.flag] = eval.value;
+                  });
+                }
+              }
+              break;
 
-          // There's been an interruption in the SSE but which has since resumed, which means the
-          // cache will have been updated with the latest values, so we can call
-          // bool variation to get the most up to date evaluation value.
+            // There's been an interruption in the SSE but which has since resumed, which means the
+            // cache will have been updated with the latest values.
+            // If we have missed any SSE events while the connection has been interrupted, we can call
+            // bool variation to get the most up to date evaluation value.
             case EventType.SSE_RESUME:
-              CfClient.getInstance().boolVariation(boolFlagName, false).then((value) {
-                print("$boolFlagName: $value");
-                setState(() {
-                  _boolFlagValue = value;
-                });
-              });
+              flagVariations();
               break;
 
             default:
               break;
           }
         }
+
         CfClient.getInstance().registerEventsListener(listener);
+        // CfClient.getInstance().destroy();
         // CfClient.getInstance().unregisterEventsListener(listener);
       }
+    });
+  }
+
+  void flagVariations() {
+    // Evaluate flag and set initial state
+    CfClient.getInstance()
+        .boolVariation(boolFlagName, false)
+        .then((variationResult) {
+      setState(() {
+        _flagValues[boolFlagName] = variationResult;
+      });
+    });
+
+    // Evaluate flag and set initial state
+    CfClient.getInstance()
+        .jsonVariation(jsonFlagName, {}).then((variationResult) {
+      setState(() {
+        _flagValues[jsonFlagName] = variationResult;
+      });
+    });
+
+    // Evaluate flag and set initial state
+    CfClient.getInstance()
+        .stringVariation(stringFlagName, "default")
+        .then((variationResult) {
+      setState(() {
+        _flagValues[stringFlagName] = variationResult;
+      });
+    });
+
+    // Evaluate flag and set initial state
+    CfClient.getInstance()
+        .numberVariation(numberFlagName, 1)
+        .then((variationResult) {
+      setState(() {
+        _flagValues[numberFlagName] = variationResult;
+      });
     });
   }
 
@@ -150,17 +171,21 @@ class _FlagState extends State<FlagState> {
             children: [
               Padding(
                 padding: const EdgeInsets.only(bottom: 16.0),
-                child: Text("$boolFlagName : $_boolFlagValue", style: const TextStyle(fontSize: 25)),
+                child: Text("$boolFlagName : ${_flagValues[boolFlagName]}",
+                    style: const TextStyle(fontSize: 25)),
               ),
               Padding(
                 padding: const EdgeInsets.only(bottom: 16.0),
-                child: Text("$stringFlagName : $_stringFlagValue", style: const TextStyle(fontSize: 25)),
+                child: Text("$stringFlagName : ${_flagValues[stringFlagName]}",
+                    style: const TextStyle(fontSize: 25)),
               ),
               Padding(
                 padding: const EdgeInsets.only(bottom: 16.0),
-                child: Text("$numberFlagName : $_numberFlagValue", style: const TextStyle(fontSize: 25)),
+                child: Text("$numberFlagName : ${_flagValues[numberFlagName]}",
+                    style: const TextStyle(fontSize: 25)),
               ),
-              Text("$jsonFlagName : $_jsonFlagValue", style: const TextStyle(fontSize: 25)),
+              Text("$jsonFlagName : ${_flagValues[jsonFlagName]}",
+                  style: const TextStyle(fontSize: 25)),
             ],
           ),
         ),